added option to build sore statically; set as default. dynamic lib is unmaintained. fix compiler warnings

+++ src/lib/sore_font.cpp
@@ -105,7 +105,7 @@ void SORE_Font::Font::Load()
     fontInfo = new FT_Byte[size];
 
     fontObj->strm().read(reinterpret_cast<char*>(fontInfo), size);
-    err = fontObj->strm().gcount();
+    err = static_cast<size_t>(fontObj->strm().gcount());
     delete fontObj;
 
     if(err != size)
